src/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.9
multidict==6.1.0
psycopg2==2.9.9
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==72.1.0
urllib3==2.2.3
wheel==0.44.0
yarl==1.11.1
